feat: restructure single compartment module and update imports
